REFACTOR : 상품 상세페이지 쿼리 오류 및 로딩 상태 적용 / 컴포넌트 분리
