+++ src/components/molcule/description-product/index.tsx
@@ -1,21 +1,20 @@
 import * as S from './style';
 import stategrey from 'assets/icons/state_grey.svg';
 import stategreen from 'assets/icons/state_green.svg';
-import { ProductDetailItem } from 'types/productType';
+import { useProduct } from 'store/product';
 
-interface DescriptionProductProps {
-  product: ProductDetailItem;
-}
+const DescriptionProduct = () => {
+  const product = useProduct();
+  if (!product) return null;
 
-const DescriptionProduct = ({ product }: DescriptionProductProps) => {
-  const circleUrl = product.product_status === '아주 좋아요' ? stategreen : stategrey;
+  const stateImgUrl = product.product_status === '아주 좋아요' ? stategreen : stategrey;
 
   return (
     <S.Container>
       <p className="name">{product.product_name}</p>
 
       <S.SubTitle>
-        <img src={circleUrl} alt="none-trade" />
+        <img src={stateImgUrl} alt="none-trade" />
 
         <p className="state">상품 상태 : {product.product_status}</p>
         <p className="time">{product.time}</p>
