+++ src/components/organism/content-product-detail/style.ts
