+++ src/pages/productPage/product-detail/index.tsx
@@ -1,25 +1,16 @@
 import {
-  BoxProductProfile,
   ButtonBack,
-  Carousel,
-  DescriptionProduct,
+  ContentProductDetail,
   FooterProductDetail,
   Header,
   KebabProductDetail,
-  Loading,
 } from 'components/index';
-import * as S from './style';
 import kebab from 'assets/icons/kebab.svg';
 import { useParams } from 'react-router-dom';
 import { useState } from 'react';
-import { useProduct } from 'store/product';
-import { useProductDetailQuery } from 'hooks/queries/products/useProductDetailQuery';
-import { Seller } from 'types/productType';
 
 const ProductDetail = () => {
   const { id } = useParams() as { id: string };
-  const { data: productDetailQuery, isLoading, isError } = useProductDetailQuery(id);
-  const product = useProduct();
   const [openKebab, setOpenKebab] = useState<boolean>(false);
 
   return (
@@ -33,33 +24,8 @@ const ProductDetail = () => {
           onClick={() => setOpenKebab(!openKebab)}
         />
       </Header>
-
       <KebabProductDetail openKebab={openKebab} id={id} />
-
-      <S.Content>
-        {isLoading && <Loading />}
-        {isError && <div>상품이 존재하지 않습니다.</div>}
-        {productDetailQuery && product && (
-          <>
-            <section className="profile">
-              <BoxProductProfile seller={product.seller as Seller} />
-            </section>
-
-            <section className="product-image">
-              <Carousel $dot="13px" $width="100%" $height="314px">
-                {product.product_image.map((url, i) => (
-                  <img src={url} alt={`img-${i}`} key={i} />
-                ))}
-              </Carousel>
-            </section>
-
-            <section className="description">
-              <DescriptionProduct product={product} />
-            </section>
-          </>
-        )}
-      </S.Content>
-
+      <ContentProductDetail id={id} />
       <FooterProductDetail />
     </>
   );
